move jetton common logic
