+++ src/benchmarks/jetton/jetton.spec.ts
@@ -2,7 +2,6 @@ import "@ton/test-utils";
 import {
     Address,
     beginCell,
-    Builder,
     Cell,
     contractAddress,
     SendMode,
@@ -11,22 +10,11 @@ import {
 import type { SandboxContract, TreasuryContract } from "@ton/sandbox";
 import { Blockchain } from "@ton/sandbox";
 
-import {
-    type Mint,
-    type ProvideWalletAddress,
-    JettonMinter,
-    storeJettonBurn,
-    storeJettonTransfer,
-    storeMint,
-} from "../contracts/output/jetton-minter-discoverable_JettonMinter";
+import { JettonMinter } from "../contracts/output/jetton-minter-discoverable_JettonMinter";
 
 import "@ton/test-utils";
 import benchmarkCodeSizeResults from "./results_code_size.json";
-import type {
-    JettonBurn,
-    JettonTransfer,
-    JettonUpdateContent,
-} from "../contracts/output/jetton-minter-discoverable_JettonMinter";
+import type { JettonUpdateContent } from "../contracts/output/jetton-minter-discoverable_JettonMinter";
 
 import {
     generateCodeSizeResults,
@@ -34,13 +22,19 @@ import {
     getStateSizeForAccount,
     getUsedGas,
     printBenchmarkTable,
-} from "../util";
+} from "../utils/gas";
 import benchmarkResults from "./results_gas.json";
 import { join, resolve } from "path";
 import { readFileSync } from "fs";
-import { storeProvideWalletAddress } from "../contracts/output/escrow_Escrow";
 import { posixNormalize } from "../../utils/filePath";
 import { type Step, writeLog } from "../../test/utils/write-vm-log";
+import {
+    getJettonWalletRaw,
+    sendBurnRaw,
+    sendDiscoveryRaw,
+    sendMintRaw,
+    sendTransferRaw,
+} from "../utils/jetton";
 
 const loadFunCJettonsBoc = () => {
     const bocMinter = readFileSync(
@@ -89,150 +83,6 @@ const deployFuncJettonMinter = async (
             sendMode: SendMode.PAY_GAS_SEPARATELY,
         }),
     };
-};
-
-const sendDiscoveryRaw = async (
-    minterAddress: Address,
-    via: SandboxContract<TreasuryContract>,
-    address: Address,
-    includeAddress: boolean,
-    value: bigint,
-) => {
-    const msg: ProvideWalletAddress = {
-        $$type: "ProvideWalletAddress",
-        queryId: 0n,
-        ownerAddress: address,
-        includeAddress: includeAddress,
-    };
-
-    const msgCell = beginCell().store(storeProvideWalletAddress(msg)).endCell();
-
-    return await via.send({
-        to: minterAddress,
-        value,
-        body: msgCell,
-        sendMode: SendMode.PAY_GAS_SEPARATELY,
-    });
-};
-
-const sendTransferRaw = async (
-    jettonWalletAddress: Address,
-    via: SandboxContract<TreasuryContract>,
-    value: bigint,
-    jetton_amount: bigint,
-    to: Address,
-    responseAddress: Address,
-    customPayload: Cell | null,
-    forward_ton_amount: bigint,
-    forwardPayload: Cell | null,
-) => {
-    const parsedForwardPayload =
-        forwardPayload != null
-            ? forwardPayload.beginParse()
-            : new Builder().storeUint(0, 1).endCell().beginParse(); //Either bit equals 0
-
-    const msg: JettonTransfer = {
-        $$type: "JettonTransfer",
-        queryId: 0n,
-        amount: jetton_amount,
-        destination: to,
-        responseDestination: responseAddress,
-        customPayload: customPayload,
-        forwardTonAmount: forward_ton_amount,
-        forwardPayload: parsedForwardPayload,
-    };
-
-    const msgCell = beginCell().store(storeJettonTransfer(msg)).endCell();
-
-    return await via.send({
-        to: jettonWalletAddress,
-        value,
-        body: msgCell,
-        sendMode: SendMode.PAY_GAS_SEPARATELY,
-    });
-};
-
-const sendMintRaw = async (
-    jettonMinterAddress: Address,
-    via: SandboxContract<TreasuryContract>,
-    to: Address,
-    jetton_amount: bigint,
-    forward_ton_amount: bigint,
-    total_ton_amount: bigint,
-) => {
-    if (total_ton_amount <= forward_ton_amount) {
-        throw new Error(
-            "Total TON amount should be greater than the forward amount",
-        );
-    }
-
-    const msg: Mint = {
-        $$type: "Mint",
-        queryId: 0n,
-        receiver: to,
-        tonAmount: total_ton_amount,
-        mintMessage: {
-            $$type: "JettonTransferInternal",
-            queryId: 0n,
-            amount: jetton_amount,
-            responseDestination: jettonMinterAddress,
-            forwardTonAmount: forward_ton_amount,
-            sender: jettonMinterAddress,
-            forwardPayload: beginCell().storeUint(0, 1).endCell().beginParse(),
-        },
-    };
-
-    const msgCell = beginCell().store(storeMint(msg)).endCell();
-
-    return await via.send({
-        to: jettonMinterAddress,
-        value: total_ton_amount + toNano("0.015"),
-        body: msgCell,
-        sendMode: SendMode.PAY_GAS_SEPARATELY,
-    });
-};
-
-const sendBurnRaw = async (
-    jettonWalletAddress: Address,
-    via: SandboxContract<TreasuryContract>,
-    value: bigint,
-    jetton_amount: bigint,
-    responseAddress: Address,
-    customPayload: Cell | null,
-) => {
-    const msg: JettonBurn = {
-        $$type: "JettonBurn",
-        queryId: 0n,
-        amount: jetton_amount,
-        responseDestination: responseAddress,
-        customPayload: customPayload,
-    };
-
-    const msgCell = beginCell().store(storeJettonBurn(msg)).endCell();
-
-    return await via.send({
-        to: jettonWalletAddress,
-        value,
-        body: msgCell,
-        sendMode: SendMode.PAY_GAS_SEPARATELY,
-    });
-};
-
-const getJettonWalletRaw = async (
-    minterAddress: Address,
-    blockchain: Blockchain,
-    walletAddress: Address,
-) => {
-    const walletAddressResult = await blockchain
-        .provider(minterAddress)
-        .get(`get_wallet_address`, [
-            {
-                type: "slice",
-                cell: beginCell().storeAddress(walletAddress).endCell(),
-            },
-        ]);
-
-    return walletAddressResult.stack.readAddress();
 };
 
 describe("Jetton", () => {
